python/requirements.txt
astroid==2.2.5
Click==7.0
colorama==0.4.1
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
pylint==2.3.1
pyodbc==4.0.26
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.4
wrapt==1.11.1
